Rename overview files and update structure
- Changed Home to Overview
- Changed Events to OverviewEvents
- Added an index.js to make importing cleaner
