+++ src/router/index.js
@@ -1,14 +1,13 @@
 import Vue from "vue";
 import VueRouter from "vue-router";
-import Home from "../views/Overview/Home.vue";
 
 Vue.use(VueRouter);
 
 const routes = [
   {
     path: "/",
-    name: "home",
-    component: Home
+    name: "overview",
+    component: () => import("@/views/Overview")
   },
   {
     path: "/access-control/local-user-management",
