+++ src/views/Overview/index.js
@@ -1,0 +1,2 @@
+import Overview from "./Overview.vue";
+export default Overview;
